Revert "refactor(api,application-generic,ws,worker,webhook): resolve circular dependencies; import refactor (#6268)"
This reverts commit 4a363ce656e933d355f1a2ed27b7586ef05bff6c.
