+++ apps/api/src/app.module.ts
@@ -59,6 +59,12 @@ const enterpriseImports = (): Array<Type | DynamicModule | Promise<DynamicModule
     }
   }
 
+  if (isClerkEnabled()) {
+    if (require('@novu/ee-auth')?.EEAuthModule) {
+      modules.push(require('@novu/ee-auth')?.EEAuthModule);
+    }
+  }
+
   return modules;
 };
