+++ apps/webhook/package.json
@@ -68,9 +68,6 @@
     "tsconfig-paths": "~4.1.0",
     "typescript": "4.9.5"
   },
-  "optionalDependencies": {
-    "@novu/ee-auth": "workspace:*"
-  },
   "workspaces": {
     "nohoist": [
       "@nestjs/platform-socket.io",
