+++ enterprise/packages/auth/package.json
@@ -15,7 +15,6 @@
     "@clerk/clerk-sdk-node": "^5.0.19",
     "@novu/dal": "workspace:*",
     "@novu/shared": "workspace:*",
-    "@novu/application-generic": "workspace:*",
     "mongoose": "^7.5.0",
     "class-transformer": "^0.5.1",
     "class-validator": "^0.14.0",
