+++ enterprise/packages/billing/package.json
@@ -47,8 +47,5 @@
     "@nestjs/swagger": "^7.1.8",
     "@nestjs/throttler": "^5.0.1",
     "@novu/dal": "workspace:*"
-  },
-  "optionalDependencies": {
-    "@novu/ee-auth": "workspace:*"
   }
 }
