+++ enterprise/packages/translation/package.json
@@ -43,8 +43,5 @@
     "@nestjs/swagger": "^7.1.8",
     "@nestjs/platform-express": "^10.2.2",
     "@novu/dal": "workspace:*"
-  },
-  "optionalDependencies": {
-    "@novu/ee-auth": "workspace:*"
   }
 }
