+++ libs/application-generic/package.json
@@ -98,6 +98,7 @@
     "slugify": "^1.4.6"
   },
   "optionalDependencies": {
+    "@novu/ee-auth": "workspace:*",
     "@novu/ee-shared-services": "workspace:*",
     "@taskforcesh/bullmq-pro": "5.1.14"
   },
