+++ libs/application-generic/src/index.ts
@@ -9,7 +9,7 @@ export * from './health/index';
 export * from './instrumentation/index';
 export * from './logging/index';
 export * from './modules';
-export * from './utils/inject-auth-providers';
+export * from './utils/inject-repositories';
 export * from './profiling';
 export * from './resilience';
 export * from './services';
