+++ libs/testing/src/ee/ee.repository.factory.ts
@@ -1,13 +1,6 @@
 import { CommunityOrganizationRepository, CommunityUserRepository, CommunityMemberRepository } from '@novu/dal';
 import { isClerkEnabled } from '@novu/shared';
 
-/**
- * We are using nx-ignore-next-line as a workaround here to avoid following circular dependency error:
- * @novu/application-generic:build --> @novu/testing:build --> @novu/ee-auth:build --> @novu/application-generic:build
- *
- * When revising EE testing, we should consider refactoring the code to potentially avoid this circular dependency.
- *
- */
 export function getEERepository<T>(className: 'OrganizationRepository' | 'MemberRepository' | 'UserRepository'): T {
   if (isClerkEnabled()) {
     switch (className) {
@@ -31,7 +24,6 @@ export function getEERepository<T>(className: 'OrganizationRepository' | 'Member
 }
 
 function getEEUserRepository() {
-  // nx-ignore-next-line
   const enterpriseModule = require('@novu/ee-auth');
   const enterpriseUserRepository = enterpriseModule?.EEUserRepository;
   const clerkClientMock = enterpriseModule?.ClerkClientMock;
@@ -40,7 +32,6 @@ function getEEUserRepository() {
 }
 
 function getEEOrganizationRepository() {
-  // nx-ignore-next-line
   const enterpriseModule = require('@novu/ee-auth');
   const enterpriseOrganizationRepository = enterpriseModule?.EEOrganizationRepository;
   const clerkClientMock = enterpriseModule?.ClerkClientMock;
@@ -49,7 +40,6 @@ function getEEOrganizationRepository() {
 }
 
 function getEEMemberRepository() {
-  // nx-ignore-next-line
   const enterpriseModule = require('@novu/ee-auth');
   const enterpriseMemberRepository = enterpriseModule?.EEMemberRepository;
   const clerkClientMock = enterpriseModule?.ClerkClientMock;
